Fix bug where double was cast to single.
Fix bug where 'RL8' variables were cast to single instead of double when
printing, causing incorrect values.

+++ cfs.c
@@ -385,7 +385,7 @@ void print_variable(CFSVariable *variable)
 		break;
 		case RL8:
 			printf("Type: RL8\n");
-			printf("Variable: %f\n", *(float *)variable->data);
+			printf("Variable: %f\n", *(double *)variable->data);
 		break;
 		case LSTR:
 			printf("Type: LSTR\n");
